Move BoreholePermissionService
