+++ src/api/Program.cs
@@ -4,6 +4,7 @@ using BDMS;
 using BDMS.Authentication;
 using BDMS.Json;
 using BDMS.Maintenance;
+using BDMS.Services;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Mvc;
